+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/fetchers/SvgRegionFetcher.kt
@@ -11,12 +11,14 @@ import com.caverock.androidsvg.PreserveAspectRatio
 import com.caverock.androidsvg.RenderOptions
 import com.caverock.androidsvg.SVG
 import com.caverock.androidsvg.SVGParseException
+import deckers.thibault.aves.channel.streams.ByteSink
 import deckers.thibault.aves.metadata.SVGParserBufferedInputStream
 import deckers.thibault.aves.metadata.SvgHelper.normalizeSize
 import deckers.thibault.aves.utils.BitmapUtils
 import deckers.thibault.aves.utils.MemoryUtils
+import deckers.thibault.aves.utils.MimeTypes
 import deckers.thibault.aves.utils.StorageUtils
-import io.flutter.plugin.common.MethodChannel
+import java.io.ByteArrayInputStream
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.concurrent.withLock
 import kotlin.math.ceil
@@ -24,14 +26,15 @@ import kotlin.math.ceil
 class SvgRegionFetcher internal constructor(
     private val context: Context,
 ) {
-    fun fetch(
+    suspend fun fetch(
         uri: Uri,
+        decoded: Boolean,
         sizeBytes: Long?,
         scale: Int,
         regionRect: Rect,
         imageWidth: Int,
         imageHeight: Int,
-        result: MethodChannel.Result,
+        result: ByteSink,
     ) {
         if (!MemoryUtils.canAllocate(sizeBytes)) {
             // opening an SVG that large would yield an OOM during parsing from `com.caverock.androidsvg.SVGParser`
@@ -79,7 +82,7 @@ class SvgRegionFetcher internal constructor(
             val targetBitmapSizeBytes = BitmapUtils.getExpectedImageSize(pixelCount.toLong(), config)
             if (!MemoryUtils.canAllocate(targetBitmapSizeBytes)) {
                 // decoding a region that large would yield an OOM when creating the bitmap
-                result.error("fetch-read-large-region", "SVG region too large for uri=$uri regionRect=$regionRect", null)
+                result.error("fetch-large-region", "SVG region too large for uri=$uri regionRect=$regionRect", null)
                 return
             }
 
@@ -88,12 +91,17 @@ class SvgRegionFetcher internal constructor(
             svg.renderToCanvas(canvas, renderOptions)
 
             bitmap = Bitmap.createBitmap(bitmap, bleedX, bleedY, targetBitmapWidth, targetBitmapHeight)
-            val bytes = BitmapUtils.getRawBytes(bitmap, recycle = true)
-            result.success(bytes)
+            val bytes = BitmapUtils.getBytes(bitmap, recycle = true, decoded = decoded, MimeTypes.SVG)
+            if (bytes == null) {
+                result.error("fetch-null", "failed to decode SVG for uri=$uri regionRect=$regionRect", null)
+            } else {
+                result.streamBytes(ByteArrayInputStream(bytes))
+                result.endOfStream()
+            }
         } catch (e: SVGParseException) {
-            result.error("fetch-parse", "failed to parse SVG for uri=$uri regionRect=$regionRect", null)
+            result.error("fetch-parse", "failed to parse SVG for uri=$uri regionRect=$regionRect", e.message)
         } catch (e: Exception) {
-            result.error("fetch-read-exception", "failed to initialize region decoder for uri=$uri regionRect=$regionRect", e.message)
+            result.error("fetch-exception", "failed to initialize region decoder for uri=$uri regionRect=$regionRect", e.message)
         }
     }