+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/fetchers/TiffRegionFetcher.kt
@@ -4,25 +4,28 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Rect
 import android.net.Uri
+import deckers.thibault.aves.channel.streams.ByteSink
 import deckers.thibault.aves.utils.BitmapUtils
-import io.flutter.plugin.common.MethodChannel
+import deckers.thibault.aves.utils.MimeTypes
 import org.beyka.tiffbitmapfactory.DecodeArea
 import org.beyka.tiffbitmapfactory.TiffBitmapFactory
+import java.io.ByteArrayInputStream
 
 class TiffRegionFetcher internal constructor(
     private val context: Context,
 ) {
-    fun fetch(
+    suspend fun fetch(
         uri: Uri,
         page: Int,
+        decoded: Boolean,
         sampleSize: Int,
         regionRect: Rect,
-        result: MethodChannel.Result,
+        result: ByteSink,
     ) {
         try {
             val pfd = context.contentResolver.openFileDescriptor(uri, "r")
             if (pfd == null) {
-                result.error("getRegion-tiff-fd", "failed to get file descriptor for uri=$uri", null)
+                result.error("fetch-fd", "failed to get file descriptor for uri=$uri", null)
                 return
             }
             pfd.use {
@@ -33,15 +36,16 @@ class TiffRegionFetcher internal constructor(
                     inDecodeArea = DecodeArea(regionRect.left, regionRect.top, regionRect.width(), regionRect.height())
                 }
                 val bitmap: Bitmap? = TiffBitmapFactory.decodeFileDescriptor(fd, options)
-                val bytes = BitmapUtils.getRawBytes(bitmap, recycle = true)
-                if (bytes != null) {
-                    result.success(bytes)
+                val bytes = BitmapUtils.getBytes(bitmap, recycle = true, decoded = decoded, MimeTypes.TIFF)
+                if (bytes == null) {
+                    result.error("fetch-null", "failed to decode region for uri=$uri page=$page regionRect=$regionRect", null)
                 } else {
-                    result.error("getRegion-tiff-null", "failed to decode region for uri=$uri page=$page regionRect=$regionRect", null)
+                    result.streamBytes(ByteArrayInputStream(bytes))
+                    result.endOfStream()
                 }
             }
         } catch (e: Exception) {
-            result.error("getRegion-tiff-read-exception", "failed to read from uri=$uri page=$page regionRect=$regionRect", e.message)
+            result.error("fetch-exception", "failed to read from uri=$uri page=$page regionRect=$regionRect", e.message)
         }
     }
 }